refactor: Delete sever component prefetching on itemspage
