+++ src/app/providers.tsx
@@ -10,6 +10,7 @@ export default function Providers({ children }: { children: ReactNode }) {
 
   return (
     <QueryClientProvider client={queryClient}>
+      <ReactQueryDevtools initialIsOpen={false} />
       <AuthProvider>{children}</AuthProvider>
     </QueryClientProvider>
   );
